manifests: add build pipelines
This makes sure we pin the versions of the tools we use to build
our images, and that we do not require the host system to have
installed any of them.

We use the same build pipeline in bouth our manifests just so
it will be reused for both runs.

Signed-off-by: Tom Gundersen <teg@jklm.no>