manifests: build fedora-31
It is the most recent release, so it makes sense. But more importantly,
it does not appear to be possible to use the ostree stages as-is on
a fedora 30 host, so when adding build pipelines, I would need to use
a fedora 31 buildpipeline to build a fedora 30 images, which just
felt off.

Signed-off-by: Tom Gundersen <teg@jklm.no>