sign-commit: add helper to sign commits
Add a small helper script that can be used to sign ostree commits.
It can be configured via a `signing.ini` file, which can contain
the path to the gpg home (`gpghome`) and the gpg key identity to
be used (`keyid`). The `signing.ini` file must stored in the same
folder as the script itself.
Additionally, sign-commit provides a command line switch called,
`--gen-keypair` that will create a new gpg key pair, with no
passphrase and create a corresponding `signing.ini` file. This is
meant to ease testing.

Example `signing.ini` file:

  [signing]
  gpghome = <path>
  keyid = <gpg key id>