osbuild-ostree: load an existing info.json
Instead of always starting with a fresh `info.json`, try to load
an existing one. Currently all the information will be overwritten
but in the near future image and commit building will be separated
and then one can only build an image on an existing commit without
attempting to rebuild the commit.