osbuild-update: add helper for ostree updates
Add a new osbuild-update helper that can be used to update ostree
based images. Initially, `osbuild-update setup` will create a
repository in the build directory, which will serve as the remote
repo to fetch updates from. It also setups up the remote entry in
the image manifest. After a commit and image have been build, the
`osbuild-update prepare` command will modify the image manifest
to include the parent commit. After building that commit, which
can of course also include other changes, via `osbuild-ostree
build --commit --sign`, a call to `osbuild-update finish` will
pull that commit into the repo so it can be served to the clients
in the image. There is a convenience `osbuild-update serve`
command that will do just that.