manifests: enable usbuard in fedora iot
Now that we a) have selinux labels in the buildroot, b) a custom
osbuild policy that will transition setfiles into the needed
setfiles_mac_t domain and 3) the CAP_MAC_ADMIN capability for the
container, it should now be possible to apply labels that are not
known to the host.
NB: the modules contained in the sub-module here need to have the
correct SELinux label. Use `relabel.sh` to do so. This of course
also needs to have osbuild > 16, with the correct labels installed.